# Conflicts:
#	Datapack
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/network/clientpackets/EnterWorld.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/network/serverpackets/AbstractItemPacket.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/network/serverpackets/UserInfo.java

+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/network/serverpackets/AbstractItemPacket.java
@@ -82,7 +82,7 @@ public abstract class AbstractItemPacket extends AbstractMaskPacket<ItemListType
 
     protected void writeItem(Item item, Player owner, WritableBuffer buffer) {
         final int mask = calculateMask(item);
-        buffer.writeByte(mask);
+        buffer.writeShort(mask);
         buffer.writeInt(item.getObjectId());
         buffer.writeInt(item.getDisplayId());
         buffer.writeByte(item.isQuestItem() || item.isEquipped() ? 0xFF : item.getLocationSlot());
